CronJob/requirements.txt
sqlalchemy 
pymysql
pandas
requests
load_dotenv
zoneinfo